renamed and created new controller route handler for get movie details info and implemented in router
