+++ src/routes/movies-api-router.ts
@@ -1,8 +1,11 @@
 import configRouter from '@utilities/config-router';
 import { router } from './common-router';
-import { getMovies } from '@controllers/fetch-movies-data';
+import { getMovieById, getMovies } from '@controllers/retrive-movies-data';
 
 // get movies collection by different type query
 router.get('/movies',getMovies);
+
+// get movie by id
+router.get('/movie',getMovieById);
 
 export default configRouter(['/api',router]);
